Feat:Actualizacion ClienteView y eliminacion de metodos de cliente de taller
