+++ SistemaTaller/src/view/ClienteView.java
@@ -1,5 +1,71 @@
 package view;
 
+import java.util.Scanner;
+
+import dao.ClienteDAO;
+import model.Cliente;
+
 public class ClienteView {
-    
+    private Scanner sc = new Scanner(System.in);
+    private ClienteDAO clienteDAO = new ClienteDAO();
+
+    public void gestionClientes() {
+
+        int opcion;
+        do {
+            System.out.println("\nGestion de clientes");
+            System.out.println("1. Agregar cliente");
+            System.out.println("2. Modificar cliente");
+            System.out.println("3. Eliminar cliente");
+            System.out.println("4. Listar clientes");
+            System.out.println("5. Volver al menu principal");
+            System.out.println("Ingrese una opcion: ");
+            opcion = sc.nextInt();
+
+            switch (opcion) {
+                case 1 -> clienteDAO.agregarCliente();
+                case 2 -> clienteDAO.modificarCliente();
+                case 3 -> clienteDAO.eliminarCliente();
+                case 4 -> clienteDAO.listarClientes();
+                case 5 -> System.out.println("Volviendo al menu principal");
+            }
+        } while (opcion != 5);
+    }
+
+    public void agregarCliente() {
+        System.out.println("Agregar cliente");
+        System.out.println("Introduce el nombre: ");
+        String nombre = sc.nextLine();
+        sc.next();
+        System.out.println("Introduce el apellido: ");
+        String apellido = sc.nextLine();
+        sc.next();
+        System.out.println("Introduce el telefono: ");
+        int telefono = sc.nextInt();
+        System.out.println("Introduce la direccion: ");
+        String direccion = sc.nextLine();
+        sc.next();
+        System.out.println("Introduce el email: ");
+        String email = sc.nextLine();
+        sc.next();
+        System.out.println("Introduce el dni: ");
+        String dni = sc.nextLine();
+        sc.next();
+
+        Cliente cliente = new Cliente(nombre, apellido, telefono, direccion, email, dni);
+        clienteDAO.listaClientes.add(cliente);
+        System.out.println("Cliente agregado correctamente");
+    }
+
+    public void eliminarCliente() {
+        Cliente cliente = clienteDAO.getClienteDni();
+        clienteDAO.listaClientes.remove(cliente);
+        System.out.println("Cliente eliminado correctamente");
+    }
+
+    public void listarClientes() {
+        for (Cliente cliente : clienteDAO.listaClientes) {
+            System.out.println(cliente);
+        }
+    }
 }
